fixup store location + add seo and subtitle
