+++ src/lib/command_runner/commands_components/EditorPreferences.svelte
@@ -1,5 +1,8 @@
 <script lang="ts">
-	import editor_preferences, { set_default_editor_preferences } from '$lib/editor_preferences';
+	import {
+		editor_preferences,
+		set_default_editor_preferences,
+	} from '$lib/stores/editor_config_store';
 	import font_preferences, { set_default_font_preferences } from '$lib/font_preferences';
 </script>
