+++ src/lib/stores/editor_config_store.ts
@@ -5,6 +5,17 @@ interface EditorConfigStore {
 	code_wrap: boolean;
 }
 
+const default_preferences = {
+	delay_function: 'throttle' as 'throttle' | 'debounce',
+	delay_duration: 300,
+};
+
+export const editor_preferences = persisted('sk__editor_preferences', default_preferences);
+
+export function set_default_editor_preferences() {
+	editor_preferences.set({ ...default_preferences });
+}
+
 const { subscribe, update } = persisted<EditorConfigStore>('editor_config', {
 	vim: false,
 	code_wrap: false,
